Add the protocol to the logs
This commit adds the protocol (HTTP or HTTPS) to the health-check logs.
This in preparation for the next commits, where a mirror can support
both HTTP and HTTPS, and health checks will be done twice: over HTTP and
HTTPS. In this case, logging that a mirror is "up" or "down" won't be
enough, we'll need to know the protocol concerned as well.

1) Changes to the `mirrorbits logs` command

Let's take a look at how `mirrorbits log MIRROR` looks like before this
commit:

```
2025-01-08 08:54:42 UTC: Mirror is down: Unreachable
2025-01-08 09:11:12 UTC: Mirror is up
```

And after this commit:

```
2025-01-08 08:55:02 UTC: HTTPS mirror is down: Unreachable
2025-01-08 09:11:12 UTC: HTTPS mirror is up
```

If you considered `mirrorbits log` as a stable format (ie. maybe you
have scripts that parse its output), then take this commit as a breaking
change. You'll need to adjust your scripts accordingly.

2) Changes to the mirrorbits stderr logs

Similarily, the mirrorbits logs (that go to stderr by default) are
updated as such.

Before this commit:

```
2025/01/09 15:35:53.121 UTC kali.darklab.sh                Down! Status: 503
2025/01/10 03:48:56.864 UTC repo.jing.rocks                Up! (509ms)
```

After this commit:

```
2025/01/09 15:35:53.121 UTC kali.darklab.sh                HTTP  Down! Status: 503
2025/01/10 03:48:56.864 UTC repo.jing.rocks                HTTPS Up! (509ms)
```

3) How to review this commit

File `mirrors/logs.go` is simple to review, there is a new field `proto`
in the struct, and updates to support this new field.

The field `proto` is of type `Protocol`, this is a new type introduced
in `mirrors/mirrors.go`.

Finally, in order to actually add this field to the log, we need to
update the signatures of `MarkMirrorUp`, `MarkMirrorDown` and
`SetMirrorState` (`mirrors/mirrors.go`), then the caller
`daemon/monitor.go` can make use of it.