Support relative URLs in the mirror field `HttpURL`
Before this commit, the value for HttpURL must start with either
`http://` or `https://`, for example `http://mirror.com/foobar/`.

With this commit, we also allow "relative URLs", ie. URLs without a
scheme, meaning that `mirror.com/foobar/` is now a valid value, for
mirrors and also for fallback mirrors (defined in mirrorbits.conf).

In case of relative URL, the protocol is assumed to be HTTP, but that's
only temporary. In upcoming commits, relative URL will mean that the
mirror supports both HTTP and HTTPS.

This change adds some complexity: all the code that uses `HttpURL` must
now check if the URL has a scheme, and deal with that. There's no
one-size-fits-all helper for that (even though for now it looks like
that, but that's because relative URL default to HTTP now, and that's
only temporary).

Additionally, codes that runs after the selection process, and that
requires a absolute URL, must now use `AbsoluteURL` instead of `HttpURL`
(that's `templates/mirrorlist.html` and `http/pagerenderer.go`).