Decouple DownReason and ExcludeReason
So far, the field ExcludeReason had two uses. First, it was set during
health-check when the mirror was found to be down (hence commited to
redis db), and it was later set during mirror selection when the mirror
was excluded (in this case, the value set was never set in the redis db,
it was just discarded after the request was served).

So these are two different things that were conflated in one field, and
that worked, but now that we're about to support both HTTP and HTTPS for
mirrors, that won't cut it.

So, this commit does two things:

1) Rename ExcludeReason to DownReason - this is the field that is set by
   the health check

2) Add a new ExcludeReason field, which is to be set by the mirror
   selection, and used by the code that comes after that and wants to
   know why the mirror was excluded

Note that, given that the redis key is also renamed, there needs to be a
database migration - that will come in a later commit.