network: Make RemoteIPFromAddr handle empty strings
`RemoteIPFromAddr` doesn't check its input, so if ever we pass it a
empty string, we get that:

```
panic: runtime error: slice bounds out of range [:-1] [recovered]
        panic: runtime error: slice bounds out of range [:-1]
```

It doesn't really happen in practice though, this function is only used
in `mirrorHandler` (`http.go`), where we pass it the field `RemoteAddr`
from the incoming request. And it seems that this field is guaranteed to
be set to `IP:port`, according to <https://pkg.go.dev/net/http#Request>.

So this commit doesn't fix any known issue, it just adds a bit of
safety, because it's easy and it doesn't hurt.